backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pandas==2.2.3
google-auth==2.35.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0
google-api-python-client==2.149.0
anthropic==0.36.0
python-dotenv==1.0.1
openpyxl==3.1.5
httpx==0.27.2
python-multipart==0.0.12
requests==2.32.3
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
tavily-python==0.5.0
